Move SafeMaskMaker
