+++ gammapy/spectrum/tests/test_make.py
@@ -6,10 +6,10 @@ import astropy.units as u
 from astropy.coordinates import Angle, SkyCoord
 from regions import CircleSkyRegion
 from gammapy.data import DataStore
+from gammapy.cube import SafeMaskMaker
 from gammapy.maps import WcsGeom, WcsNDMap
 from gammapy.spectrum import (
     ReflectedRegionsBackgroundMaker,
-    SafeMaskMaker,
     SpectrumDatasetMaker,
 )
 from gammapy.utils.testing import assert_quantity_allclose, requires_data
